requirements-flask.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.4.0

# Vector database
qdrant-client==1.9.2

# Google APIs
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.140.0
google-generativeai==0.8.1

# Document processing
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5

# Data processing
pandas==2.2.2
numpy==1.26.4

# Lightweight ML dependencies (CPU-only torch for Docker)
torch==2.4.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.19.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.4.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html

# Sentence transformers with minimal dependencies
sentence-transformers==3.0.1
transformers==4.44.2
tokenizers==0.19.1
huggingface-hub==0.24.6

# Additional utilities
requests==2.32.3
aiofiles==24.1.0
httpx==0.27.0

# Scheduling for sync
schedule==1.2.2

# File monitoring
watchdog==5.0.0

# Logging
loguru==0.7.2

# Production server
gunicorn==21.2.0

langchain
langchain-qdrant
langchain-community
